将链接打开行为移入浏览导航
