+++ src/components/Settings/PluginComponentsAndPages/General/General.vue
@@ -33,48 +33,6 @@ const langOptions = computed(() => {
   ]
 })
 
-const openModeOptions = computed(() => {
-  return [
-    {
-      label: t('settings.link_opening_behavior_opt.current_tab'),
-      value: 'currentTab',
-    },
-    {
-      label: t('settings.link_opening_behavior_opt.current_tab_if_not_homepage'),
-      value: 'currentTabIfNotHomepage',
-    },
-    {
-      label: t('settings.link_opening_behavior_opt.background'),
-      value: 'background',
-    },
-    {
-      label: t('settings.link_opening_behavior_opt.new_tab'),
-      value: 'newTab',
-    },
-  ]
-})
-
-const videoCardOpenModeOptions = computed(() => {
-  return [
-    {
-      label: t('settings.link_opening_behavior_opt.current_tab'),
-      value: 'currentTab',
-    },
-    {
-      label: t('settings.link_opening_behavior_opt.drawer'),
-      value: 'drawer',
-    },
-    {
-      label: t('settings.link_opening_behavior_opt.background'),
-      value: 'background',
-    },
-    {
-      label: t('settings.link_opening_behavior_opt.new_tab'),
-      value: 'newTab',
-    },
-  ]
-})
-
 watch(() => settings.value.language, (newValue) => {
   locale.value = newValue
 })
@@ -109,32 +67,6 @@ watch(() => settings.value.language, (newValue) => {
 
       <SettingsItem :title="$t('settings.enable_horizontal_scrolling')" :desc="$t('settings.enable_horizontal_scrolling_desc')" right-width="auto">
         <Radio v-model="settings.enableHorizontalScrolling" />
-      </SettingsItem>
-    </SettingsItemGroup>
-
-    <SettingsItemGroup :title="$t('settings.group_link_opening_behavior')">
-      <SettingsItem :title="$t('settings.top_bar_link_opening_behavior')" :desc="$t('settings.link_opening_behavior_desc')" right-width="auto">
-        <Select v-model="settings.topBarLinkOpenMode" :options="openModeOptions" w="160px" />
-      </SettingsItem>
-      <SettingsItem :title="$t('settings.video_card_link_opening_behavior')" :desc="$t('settings.video_card_link_opening_behavior_desc')" right-width="auto">
-        <Select
-          v-model="settings.videoCardLinkOpenMode"
-          :options="videoCardOpenModeOptions"
-          w="160px"
-        />
-      </SettingsItem>
-      <SettingsItem :title="$t('settings.search_bar_link_opening_behavior')" :desc="$t('settings.link_opening_behavior_desc')" right-width="auto">
-        <Select
-          v-model="settings.searchBarLinkOpenMode"
-          :options="openModeOptions"
-          w="160px"
-        />
-      </SettingsItem>
-      <SettingsItem right-width="auto">
-        <template #title>
-          <div v-html="$t('settings.close_drawer_without_pressing_esc_again')" />
-        </template>
-        <Radio v-model="settings.closeDrawerWithoutPressingEscAgain" />
       </SettingsItem>
     </SettingsItemGroup>
