+++ src/components/Settings/Settings.vue
@@ -135,7 +135,13 @@ const settingsSearchEntries: SettingsSearchEntry[] = [
   { titleKey: 'settings.select_language', menu: MenuType.General, targetTitleKey: 'settings.group_language' },
   { titleKey: 'settings.group_interaction_layout', menu: MenuType.General, targetTitleKey: 'settings.group_interaction_layout' },
   { titleKey: 'settings.touch_screen_optimization', menu: MenuType.General, targetTitleKey: 'settings.group_interaction_layout' },
-  { titleKey: 'settings.group_link_opening_behavior', menu: MenuType.General, targetTitleKey: 'settings.group_link_opening_behavior' },
+  {
+    titleKey: 'settings.group_link_opening_behavior',
+    menu: MenuType.Navigation,
+    targetTitleKey: 'settings.group_link_opening_behavior',
+    secondaryPage: 'link-opening',
+    secondaryStorageKey: 'bewly-settings-navigation-page',
+  },
   { titleKey: 'settings.group_ad_blocking', menu: MenuType.General, targetTitleKey: 'settings.group_ad_blocking' },
   { titleKey: 'settings.block_ads', menu: MenuType.General, targetTitleKey: 'settings.group_ad_blocking' },
   { titleKey: 'settings.group_clean_share_link', menu: MenuType.General, targetTitleKey: 'settings.group_clean_share_link' },
